Pass SERVICE to filter newSession
As the MXS_SESSION no longer identifies which service the filter belongs
to, the parameter must be given to the newSession API function. With this,
the mxs::Reply constructor can be made explicit so that implicit
construction from null pointers is no longer done. With this change, the
session_set_reply now returns the correct origin of all generated queries.